refactor: convert get_boundingbox_from_kml from command to function and remove unnecessary lines
